refactor: split file list tests and graph rendering
